fix(gust): timer.sleep ticks u64->u32 — dissolves clean (both exports, 1-atom TCB)
synth's ARM backend loud-declines a u64 param in a frame-backing fn that makes a
call; v1 bounds ticks < 2^31 which fits u32, so the seam takes u32 ticks (widened
to u64 internally for the deadline). synth 0.49.0 now compiles BOTH sleep+slept
(1783 B, undefined = read32 only, executor fns inlined). Provider dissolve complete.

+++ benches/gust/drivers/timer-provider/src/lib.rs
@@ -75,11 +75,14 @@ impl Guest for P {
     /// `set_deadline_sets_only_h`: a Free/Done/out-of-range handle is a
     /// no-op) — so this marshalling layer re-implements no admission
     /// decision, only the wrap-safe arithmetic and the ticks-range guard.
-    fn sleep(handle: u32, ticks: u64) -> u32 {
-        if ticks >= (1u64 << 31) {
+    fn sleep(handle: u32, ticks: u32) -> u32 {
+        // ticks is u32 at the seam (v1 bound < 2^31 fits u32) — a u64 seam param
+        // makes synth's ARM backend loud-decline the frame-backing sleep (u64 param
+        // + a call); widening internally keeps the deadline math full-width.
+        if ticks >= (1u32 << 31) {
             return 0xFFFF_FFFF;
         }
-        let d = now().wrapping_add(ticks);
+        let d = now().wrapping_add(ticks as u64);
         let t = unsafe { tasks() };
         t.set_deadline(handle, d);
         0
